seperated main class in a separate file
